add: upgrade profile separate file
